TSC: migrated pipes
