--- src/pipes/filtering/index.js
@@ -1,4 +1,0 @@
-export * from './lowpassFilter';
-export * from './highpassFilter';
-export * from './bandpassFilter';
-export * from './notchFilter';
