+++ src/pipes/filtering/index.ts
@@ -1,0 +1,4 @@
+export * from "./lowpassFilter";
+export * from "./highpassFilter";
+export * from "./bandpassFilter";
+export * from "./notchFilter";
