+++ src/pipes/frequency/betaPower.ts
