+++ src/pipes/frequency/deltaPower.ts
