+++ src/pipes/frequency/gammaPower.ts
