--- src/pipes/frequency/index.js
@@ -1,10 +1,0 @@
-export * from './alphaPower';
-export * from './averagePower';
-export * from './betaPower';
-export * from './bufferFFT';
-export * from './fft';
-export * from './deltaPower';
-export * from './gammaPower';
-export * from './powerByBand';
-export * from './sliceFFT';
-export * from './thetaPower';
