+++ src/pipes/frequency/powerByBand.ts
