+++ src/pipes/frequency/thetaPower.ts
