--- src/pipes/reducer/index.js
@@ -1,1 +1,0 @@
-export * from "./averageDeep";
