+++ src/pipes/utility/removeChannels.ts
